scrapers/website_scraper.py
import aiohttp
from bs4 import BeautifulSoup
from typing import Dict, Any
from config.config import Config
from .base_scraper import BaseScraper
from utils.logger import setup_logger

logger = setup_logger()


class WebsiteScraper(BaseScraper):
    def __init__(self):
        super().__init__()
        self.urls = Config.WEBSITE_URLS

    async def fetch(self) -> Dict[str, Any]:
        """Fetch website content"""
        results = {}
        async with aiohttp.ClientSession() as session:
            for section, url in self.urls.items():
                try:
                    async with session.get(url) as response:
                        if response.status == 200:
                            html = await response.text()
                            results[section] = self._parse_section(html, section)
                except Exception as e:
                    logger.error(f"Error fetching {section}: {e}")
        return results

    def _parse_section(self, html: str, section: str) -> str:
        """Parse specific section content"""
        # Placeholder implementation
        return f"Content for {section}"

    async def validate(self, data: Dict[str, Any]) -> bool:
        """Validate website data"""
        return all(isinstance(v, str) for v in data.values())
