requirements.txt
# Core dependencies
python-telegram-bot==20.3
openai==0.27.8
python-dotenv==1.0.0

# Database
pymongo==4.4.1
redis==4.6.0

# Web scraping
aiohttp==3.8.5
beautifulsoup4==4.12.2
PyPDF2==3.0.1

# Async support
# asyncio==3.4.3
aiofiles==23.1.0
motor==3.1.1

# Utilities
python-dateutil==2.8.2
validators==0.20.0

# Logging and monitoring

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0
